A set of Progress Dialog changes. Sorry for the large commit, but all these changes rolled into each other.
Change the Progress Dialog to use Strings instead of Resource IDs. (This breaks the current API)
Added getCurrentDialog() to both BetterListActivity and BetterDefaultActivity. This can be used to get the currently displaying Dialog.
Also added a constant DIALOG_PROGRESS (0) used by onCreateDialog when dialogs are created. Perhaps change this constant to avoid clashes.
Added a onPrepareDialog method to ensure the title and message of the dialog is always upto date and correct.
Changed setProgressDialog{Title,Msg} to update the current displayed Progress Dialog if it exists. I'm unsure if these changes are best, but this fix the behaviour Donn Felker was expecting.