Changed ListAdapterWithProgress to not cache the Progress view.
This is on advise from a Google IO talk which says do not cache any Views from
a List. Now the progress View is created inside the getView method, or reuses
the convertView. This simpifies the constructor and no longer requires the
ListView be passed in. This also has the side effect we now have to view types
ITEM_VIEW_TYPE_PROGRESS and ITEM_VIEW_TYPE_NORMAL. A Adapter who overrides
either getItemViewType or getViewTypeCount must be aware of this.